+++ tests/unit/src/kakarot/accounts/eoa/test_library.py
@@ -75,39 +75,3 @@ class TestLibrary:
                 calls, list(calldata)
             ).call()
         ).result.response == expected_result
-
-    async def test_approval_is_infinite(
-        self, externally_owned_account, mock_kakarot, eth
-    ):
-        # Ensure approval is max after deployment
-        assert (
-            str(
-                (
-                    await eth.allowance(
-                        externally_owned_account.contract_address,
-                        mock_kakarot.contract_address,
-                    ).call()
-                ).result.remaining
-            )
-            == "Uint256(low=340282366920938463463374607431768211455, high=340282366920938463463374607431768211455)"
-        )
-
-        # Perform a token transfer
-        mock_kakarot.transfer_from_contract(
-            externally_owned_account.contract_address,
-            mock_kakarot.contract_address,
-            (1000, 0),
-        )
-
-        # After a token transfer by kakarot, the allowance should still be max
-        assert (
-            str(
-                (
-                    await eth.allowance(
-                        externally_owned_account.contract_address,
-                        mock_kakarot.contract_address,
-                    ).call()
-                ).result.remaining
-            )
-            == "Uint256(low=340282366920938463463374607431768211455, high=340282366920938463463374607431768211455)"
-        )
